pyproject.toml
[project]
name = "lablink-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Shampad Sharkar",email = "shampad@anymindgroup.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=6.0.2,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "gunicorn (>=25.0.3,<26.0.0)",
    "drf-yasg (>=1.21.14,<2.0.0)",
    "pillow (>=12.1.0,<13.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-environ (>=0.11.2,<1.0.0)",
    "environ (>=1.0,<2.0)",
    "django-env (>=0.2.1,<0.3.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
